custom exceptions

+++ app/Controllers/Http/TodosController.ts
@@ -7,12 +7,16 @@ import fs from 'fs'
 import UpdateTodoStatusValidator from 'App/Validators/UpdateTodoStatusValidator';
 import Database from '@ioc:Adonis/Lucid/Database';
 import User from 'App/Models/User';
+import BadRequestException from '../../Exceptions/BadRequestException';
+import UnsupportedMediaTypeException from '../../Exceptions/UnsupportedMediaTypeException';
+import UnAuthorizedException from '../../Exceptions/UnAuthorizedException';
+import ForbiddenException from '../../Exceptions/ForbiddenException';
 
 export default class TodosController {
     // 본인이 작성한 Todo 리스트 조회
     async list( { auth, response }: HttpContextContract ) {
         if (!auth.user) {
-            return response.status(401).send('Unauthorized')   
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
         } 
         
         const user = await User.query().preload('todos').where('id', auth.user.id).first()
@@ -20,10 +24,8 @@ export default class TodosController {
         if (user) {
             return user.todos
         } else {
-            return response.status(401).send('Unauthorized')
-        }
-
-        
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
+        }        
     }
 
     // Todo 생성
@@ -34,12 +36,12 @@ export default class TodosController {
         const { content, status } = validatedData;
 
         if (!auth.user) {
-            return response.status(401).send('Unauthorized')
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
         }
         
+        const user = auth.user
+
         try {
-            const user = auth.user
-        
             // New Todo
             const newTodo = new Todo()
             const data = {
@@ -62,7 +64,7 @@ export default class TodosController {
 
         } catch(error) {
             await trx.rollback()
-            return response.send(error.message)
+            throw new Error(error.message)
         }
     }
 
@@ -72,7 +74,8 @@ export default class TodosController {
         const todo = await Todo.find(id)
 
         if (!todo) {
-            return response.status(400).send('BAD REQUEST')
+            throw new BadRequestException('잘못된 요청입니다.', 400)
+            // return response.status(400).send('BAD REQUEST')
         }
         
         return response.status(200).send(todo)
@@ -89,17 +92,19 @@ export default class TodosController {
         
         // id에 해당하는 todo가 없을 때
         if (!todo) {
-            return response.status(400).send('BAD REQUEST')
+            throw new BadRequestException('잘못된 요청입니다.', 400)
+            // return response.status(400).send('BAD REQUEST')
         }
 
         // auth.user 가 존재하지 않을 때
         if (!auth.user) {
-            return response.status(401).send('Unauthorized')
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
         }
 
         // 본인이 작성한 todo가 아닐 때
         if (auth.user.id !== todo.userId) {
-            return response.status(403).send('본인이 작성한 Todo만 수정 가능합니다.')
+            throw new ForbiddenException('본인의 Todo만 수정 가능합니다.', 403)
+            // return response.status(403).send('본인이 작성한 Todo만 수정 가능합니다.')
         }
 
         try {
@@ -112,7 +117,7 @@ export default class TodosController {
 
         } catch (error) {
             await trx.rollback()
-            return response.send(error.message)
+            throw new Error(error.message)
         }
     }
 
@@ -120,23 +125,36 @@ export default class TodosController {
     async delete( { auth, request, response }: HttpContextContract ) {
         const trx = await Database.transaction()
 
+        if (!auth.user) {
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
+        }
+        
+        if (!request.param('id') || !Number(request.param('id'))) {
+            throw new BadRequestException('잘못된 요청입니다.', 400)
+        }
+
+        const id: number = request.param('id')
+        const todo = await Todo.query().preload('user').where('id', id).first()
+
+        if (!todo) {
+            throw new BadRequestException('잘못된 요청입니다.', 400)
+        }
+
+        if (todo.user.id !== auth.user?.id) {
+            throw new ForbiddenException('본인의 Todo만 삭제 가능합니다.', 403)
+            // return response.status(403).send('본인의 Todo만 삭제할 수 있습니다.')
+        }
+
         try {
-            const id: number = request.param('id')
-            const [ todo ] = await Todo.query().preload('user').where('id', id)
-
-            if (todo.user.id !== auth.user?.id) {
-                return response.status(403).send('본인의 Todo만 삭제할 수 있습니다.')
-            }
-
             todo.useTransaction(trx)
             await todo.delete()
             await trx.commit()
 
-            return response.status(204).send(`todo:${id}, successfully deleted`)
+            return response.status(204).send(`todo:${id}, successfully Deleted`)
 
         } catch(error) {
             await trx.rollback()
-            return response.status(400).send('BAD REQUEST')
+            throw new Error(error.message)
         }
     }
 
@@ -144,7 +162,7 @@ export default class TodosController {
     async upload( { auth, request, response }: HttpContextContract) {
 
         if (!auth.user) {
-            return response.status(401).send('Unauthorized')
+            throw new UnAuthorizedException('승인되지 않은 사용자입니다.', 401)
         }   
 
         const userId = auth.user?.id
@@ -155,7 +173,8 @@ export default class TodosController {
 
         // file validator의 extname은 왜 안먹는지 모르겠다.
         if (file.extname !== 'csv') {
-            return response.status(415).send('Extname of file should only be \'csv\'')
+            throw new UnsupportedMediaTypeException('확장자가 CSV인 파일만 업로드 가능합니다.', 415)
+            // return response.status(415).send('Extname of file should only be \'csv\'')
         }
 
         try {
@@ -170,7 +189,7 @@ export default class TodosController {
 
             // 첫 row가 content가 아닐 경우, 오류 발생 
             if (menuName !== 'content') {
-                throw SyntaxError('The first row\'s name should only be the \'content\'')
+                throw new BadRequestException('잘못 작성된 파일입니다.', 400)
             }
             
             // TODO: 확인해봐야함
@@ -202,7 +221,7 @@ export default class TodosController {
         } catch(error) {
             await trx.rollback()
             console.log(error)
-            return response.status(400).send(error.message)
+            throw new BadRequestException('잘못된 요청입니다.', 400)
         }
     }
 }
